refactor: Extract version utilities to separate module
Reduce cli.js file size by extracting version-related functions
to a new version.js module. This brings cli.js under the 1000 line limit.
